Integration tests for S2S Auth
