+++ build.gradle
@@ -205,6 +205,9 @@ dependencies {
     compileOnly group: 'org.projectlombok', name: 'lombok', version: '1.18.6'
     testCompileOnly group: 'org.projectlombok', name: 'lombok', version: '1.18.6'
     testAnnotationProcessor group: 'org.projectlombok', name: 'lombok', version: '1.18.6'
+    integrationTestCompile group: 'org.projectlombok', name: 'lombok', version: '1.18.6'
+    integrationTestAnnotationProcessor group: 'org.projectlombok', name: 'lombok', version: '1.18.6'
+    testCompile "com.github.tomakehurst:wiremock:2.19.0"
     compile group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: '2.9.9.1'
 
     compile group: 'org.springframework.boot', name: 'spring-boot-starter-web', version: versions.springBoot
