+++ charts/am-lib-test/Chart.yaml
@@ -2,3 +2,5 @@ name: am-lib-test
 version: 0.0.1
 description: Helm chart for the HMCTS Access Management testing service
 home: https://github.com/hmcts/am-lib
+maintainers:
+  - name: HMCTS Access Management Team
