+++ src/integrationTest/resources/application-test.yml
@@ -7,3 +7,10 @@ spring:
   profiles: test
   flyway:
     enabled: false
+  resources:
+    add-mappings: false
+  mvc:
+    throw-exception-if-no-handler-found: true
+security:
+  authorisedServices:
+    - "am"
